Re-build AST generations script
Made a number of changes to the API that the AST presents as as result.

We don't use const as much as we should anymore, but given that I'm
working off of someone else's design in pieces without a complete
understanding of the entire design I think this is for the best.

New AST script relies on using classes to define what is generated
rather than parsing the strings. Slightly less easy to make additions
but should be much more flexible for adding specific behavior to
different nodes of the AST as needed